Move models on BackupBloc to backup_model.dart
