+++ lib/bloc/backup/backup_actions.dart
@@ -1,5 +1,4 @@
 import 'package:breez/bloc/async_action.dart';
-import 'package:breez/bloc/backup/backup_bloc.dart';
 import 'package:breez/bloc/backup/backup_model.dart';
 
 class SaveBackupKey extends AsyncAction {
