+++ lib/routes/initial_walkthrough/dialogs/widgets/snapshot_info_tile.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:archive/archive_io.dart';
 import 'package:breez/bloc/backup/backup_actions.dart';
 import 'package:breez/bloc/backup/backup_bloc.dart';
+import 'package:breez/bloc/backup/backup_model.dart';
 import 'package:breez/bloc/blocs_provider.dart';
 import 'package:breez/services/breezlib/data/messages.pbgrpc.dart';
 import 'package:breez/utils/date.dart';
